Updated Image to first download the image with a loading indicator before loading
This has the side-effect of adding 500-700ms of load time per image. This adds up over time as more images are loaded.

While it appears to be faster, it feels slower when viewing individual images.

There's most-likely be a better way of loading images into a canvas with a progress indicator that doesn't add lag.