Added code to force deletion if it fails
While there's a native `deleteOnFail` boolean, this only seems to work on MacOS. To fix this, Node.js can do it instead.

This occurs when trying to delete from a Windows fileshare. When doing so, it can't move to the recycle bin because that's just not how Windows works. Now, it will force the folder to be permanently deleted.